+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Update Android SDK to 5.9.0
 * Update iOS SDK to 3.2.0
 * Added new method, setDMAParamsForEEA(), for setting DMA compliance parameters.
+* Removed the deprecated setDebug() method
 
 6.0.0 Dec 1, 2023
 * Update Android SDK to 5.7.5